requirements.txt
# =======================
# ✅ Core Streamlit
# =======================
streamlit>=1.31.0

# =======================
# ✅ UI Enhancement
# =======================
streamlit-option-menu>=0.3.6
streamlit-elements>=0.1.0

# =======================
# ✅ Data Handling
# =======================
pandas>=1.5.3
numpy>=1.23.5

# =======================
# ✅ Text Processing & NLP
# =======================
nltk>=3.8.1
Sastrawi>=1.0.1
scikit-learn>=1.3.0
scipy>=1.10.0

# =======================
# ✅ Plotting & Visualization
# =======================
plotly>=5.18.0
seaborn>=0.12.2

# =======================
# ✅ LLM / Google Gemini API
# =======================
google-generativeai>=0.3.2

# =======================
# ✅ PDF Generation
# =======================
fpdf2>=2.7.6

# =======================
# ✅ Firebase Integration
# =======================
firebase-admin>=6.4.0

# =======================
# ✅ Utilities
# =======================
requests>=2.31.0
pytz>=2024.1
